@@ -30,12 +30,13 @@
     "cldr-dates-full": "^39.0.0",
     "cldr-numbers-full": "^39.0.0",
     "file-saver": "^2.0.5",
+    "flexlayout-react": "^0.5.12",
     "i18next": "^20.4.0",
     "jquery": "^3.5.1",
     "merge": "^1.2.1",
     "oidc-client": "^1.9.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
     "react-i18next": "^11.11.4",
     "react-router-bootstrap": "^0.25.0",
     "react-router-dom": "^5.1.2",
